# swagger.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Oganesyan_WebAPI",
        "version": "v1"
    },
    "paths": {
...
        "/api/Solutions/add": {
            "post": {
                "tags": [
                    "Solutions"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolutionCreateDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolutionCreateDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolutionCreateDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Solution"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Solution"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Solution"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Solutions/{id}": {
            "get": {
                "tags": [
                    "Solutions"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Solution"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Solution"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Solution"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/Users/add": {
            "post": {
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/Users/update/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/Users/update": {
            "put": {
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/Users/change/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/Users/delete/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Exercise": {
                "required": [
                    "correctAnswer",
                    "difficulty",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "maxLength": 200,
                        "minLength": 1,
                        "type": "string"
                    },
                    "difficulty": {
                        "$ref": "#/components/schemas/ExerciseDifficulty"
                    },
                    "correctAnswer": {
                        "minLength": 1,
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
...
            "ExerciseDifficulty": {
                "enum": [
                    0,
                    1,
                    2
                ],
                "type": "integer",
                "format": "int32"
            },
...
            "Solution": {
                "required": [
                    "exerciseId",
                    "userAnswer"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "exerciseId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "exercise": {
                        "$ref": "#/components/schemas/Exercise"
                    },
                    "userAnswer": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "isCorrect": {
                        "type": "boolean"
                    },
                    "submittedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "result": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "SolutionCreateDto": {
                "required": [
                    "exerciseId",
                    "userAnswer"
                ],
                "type": "object",
                "properties": {
                    "exerciseId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "userAnswer": {
                        "minLength": 1,
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
            "UserCreateDto": {
                "required": [
                    "login",
                    "passwordHash",
                    "userName"
                ],
                "type": "object",
                "properties": {
                    "userName": {
                        "maxLength": 50,
                        "minLength": 1,
                        "type": "string"
                    },
                    "login": {
                        "maxLength": 50,
                        "minLength": 1,
                        "type": "string"
                    },
                    "passwordHash": {
                        "minLength": 1,
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
            "UserDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "userName": {
                        "type": "string",
                        "nullable": true
                    },
                    "login": {
                        "type": "string",
                        "nullable": true
                    },
                    "isAdmin": {
                        "type": "boolean"
                    }
                },
                "additionalProperties": false
            },
...
            "UserUpdateDto": {
                "type": "object",
                "properties": {
                    "userName": {
                        "maxLength": 50,
                        "type": "string",
                        "nullable": true
                    },
                    "passwordHash": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Вставьте JWT токен",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}